Radio messages

+++ bacon/res/shaders/stalker.frag
@@ -22,8 +22,8 @@ void main() {
 		vec2 randCoords = vec2(gl_FragCoord[0], gl_FragCoord[1]);
 		float chunk = 2.0;
 		chunk += rand(vec2(v_texCoords[1], u_elapsed)) * 4.0;
-		randCoords[0] -= mod(int(randCoords[0]), int(chunk));
-		randCoords[1] -= mod(int(randCoords[1]), 2);
+		randCoords[0] -= mod(float(randCoords[0]), float(chunk));
+		randCoords[1] -= mod(float(randCoords[1]), 2.0);
 
 		float r = (u_elapsed * 4.0) + texColor[0];
 		r += rand(randCoords);
